Added the module prefix to the function names and moved the buffer capacity define to header.
This was done because I'm a dumbass and forgot the prefix, and the buffer capacity was
moved to properly handle read, since data need to be able to handle size up to this value.

I also removed partial read capacity from the module, since there is no need and it just
introduces potential problem of partial read.